requirements.txt
benepar==0.2.0
compound-split
levenshtein
scikit-learn
sacremoses
protobuf==3.20.0
matplotlib
xlrd
nltk==3.5
overrides==3.1.0
pandas==1.0.3
pytorch_lightning==1.7.7
requests==2.23.0
tqdm==4.48.0
transformers==4.25.1
numpy==1.23.1
